# Backend/LARAVEL-B2B-TRAVEL-APP/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel B2B Travel Application Documentations",
        "description": "An API of cool stuffs",
        "contact": {
            "name": "Travel App",
            "url": "https://travel_app",
            "email": "johnckaris@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local server"
        },
        {
            "url": "http://test.travel.com/quicksava_test_api",
            "description": "Staging server"
        },
        {
            "url": "https://api.travel.com",
            "description": "Production server"
        }
    ],
    "paths": {
        "/api/accommodations": {
            "get": {
                "tags": [
                    "Accommodations"
                ],
                "summary": "Get all accommodations",
                "operationId": "a169711a099a14008d28efdeb79032c1",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/accommodations/{id}": {
            "get": {
                "tags": [
                    "Accommodations"
                ],
                "summary": "Get an accommodation by ID",
                "operationId": "0565bd00d9cadd96866b092b83328237",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the accommodation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Accommodation not found"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Authentications"
                ],
                "summary": "Register a new user",
                "operationId": "73fd9b61dff4256cc0ee710de66723d9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentications"
                ],
                "summary": "Authenticate user and generate access token",
                "operationId": "bb5bdca9f30cb60a33c1ae936a833da2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentications"
                ],
                "summary": "Revoke user's access tokens and log them out",
                "operationId": "618f033d5ecfff68b1d0880a5f399d7b",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Accommodations",
            "description": "Operations related to user authentication"
        },
        {
            "name": "Authentications",
            "description": "Endpoints for user registration, login, and logout"
        }
    ]
}